+++ packages/database/drizzle/0017_dashing_randall_flagg.sql
@@ -1,0 +1,1 @@
+ALTER TYPE "public"."symbol_type" ADD VALUE 'hyperliquid_vault';